# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shradha Oil Center ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/expiring:
    get:
      operationId: getExpiringSubscriptions
      tags: [subscriptions]
      summary: List subscriptions expiring within a window
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Expiring subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriptionListItem"
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all subscriptions with company details
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriptionListItem"
  /subscriptions/create:
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a company and its subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListItem"
  /subscriptions/{id}/renew:
    put:
      operationId: renewSubscription
      tags: [subscriptions]
      summary: Renew a subscription (extend by plan period, mark paid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListItem"
  /subscriptions/{id}/plan:
    put:
      operationId: changeSubscriptionPlan
      tags: [subscriptions]
      summary: Upgrade or downgrade the subscription plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePlanInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListItem"
  /subscriptions/{id}/suspend:
    put:
      operationId: suspendSubscription
      tags: [subscriptions]
      summary: Suspend a subscription account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListItem"
  /subscriptions/{id}/activate:
    put:
      operationId: activateSubscription
      tags: [subscriptions]
      summary: Re-activate a suspended subscription account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListItem"
  /subscriptions/{id}:
    put:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Edit a tenant company and its subscription details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionListItem"
        "400":
          description: Invalid input
        "404":
          description: Subscription not found
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Permanently delete a tenant company, its subscription and all its data (destructive)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSubscriptionResult"
        "404":
          description: Subscription not found
components:
  schemas:
    SubscriptionListItem:
      type: object
      required: [id, companyId, companyName, planName, subscriptionStartDate, subscriptionEndDate, subscriptionAmount, paymentStatus, subscriptionStatus, daysRemaining]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        ownerName:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        planName:
          type: string
        subscriptionStartDate:
          type: string
        subscriptionEndDate:
          type: string
        subscriptionAmount:
          type: number
        paymentStatus:
          type: string
          enum: [paid, pending, overdue]
        subscriptionStatus:
          type: string
          enum: [active, expired, suspended]
        lastPaymentDate:
          type: ["string", "null"]
        nextDueDate:
          type: ["string", "null"]
        daysRemaining:
          type: integer
...
    CreateSubscriptionInput:
      type: object
      required: [companyName, planName, subscriptionAmount, subscriptionStartDate]
      properties:
        companyName:
          type: string
        ownerName:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        planName:
          type: string
          enum: [monthly, quarterly, half_yearly, yearly]
        subscriptionAmount:
          type: number
        subscriptionStartDate:
          type: string
        paymentStatus:
          type: string
          enum: [paid, pending, overdue]
        adminUsername:
          type: ["string", "null"]
          description: >-
            Optional login username for the new company's admin account. When
            provided together with adminPassword, an active admin user scoped to
            the new company is created so the company can log in immediately.
        adminPassword:
          type: ["string", "null"]
          description: Login password for the new company's admin account.
    ChangePlanInput:
      type: object
      required: [planName, subscriptionAmount]
      properties:
        planName:
          type: string
          enum: [monthly, quarterly, half_yearly, yearly]
        subscriptionAmount:
          type: number
    UpdateSubscriptionInput:
      type: object
      required: [companyName, planName, subscriptionAmount, subscriptionStartDate, subscriptionEndDate, paymentStatus]
      properties:
        companyName:
          type: string
        ownerName:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        planName:
          type: string
          enum: [monthly, quarterly, half_yearly, yearly]
        subscriptionAmount:
          type: number
        subscriptionStartDate:
          type: string
        subscriptionEndDate:
          type: string
        paymentStatus:
          type: string
          enum: [paid, pending, overdue]
    DeleteSubscriptionResult:
      type: object
      required: [message, companyName]
      properties:
        message:
          type: string
        companyName:
          type: string
...